# application/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "API for managing a mechanic shop, including customers, employees, service tickets, inventory, and services"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions: 
  bearerAuth: # name of the security scheme
    type: apiKey
    name: Authorization
    in: header
    description: "JWT token, prefixed with 'Bearer '"
parameters:
...
  NameParam:
    name: name
    in: query
    type: string
    required: false
...
paths:
...
  /customers/login:
    post:
      tags:
        - Customer
      summary: "Customer login (generates token)"
      description: "Logs in a customer using email and password. generates and return a customer specific token. Rate limited to 3 requests per 10 minutes."
      parameters:
        - in: body
          name: body
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginSuccessResponse"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        429:
          description: "Too Many Requests"
          produces:
            - text/html
          schema:
            type: string
            example: "<!doctype html>\n<html lang=en>\n<title>429 Too Many Requests</title>\n<h1>Too Many Requests</h1>\n<p>3 per 10 minute</p>"
          examples:
            text/html: |
              <!doctype html>
              <html lang=en>
              <title>429 Too Many Requests</title>
              <h1>Too Many Requests</h1>
              <p>3 per 10 minute</p>
  /customers:
    post:
      tags:
        - Customer
      summary: "Create new customer"
      description: "Registers a new customer. Rate limited to 3 requests per hour."
      parameters:
        - in: body
          name: customer
          description: "Information required to create a new Customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        409:
          description: "Email already exists"
          schema:
            $ref: "#/definitions/CustomerConflictResponse"
        429:
          description: "Too Many Requests"
          produces:
            - text/html
          schema:
            type: string
          examples:
            text/html: |
              <!doctype html>
              <html lang=en>
              <title>429 Too Many Requests</title>
              <h1>Too Many Requests</h1>
              <p>3 per 1 hour</p>
        500:
          $ref: "#/definitions/ServerErrorResponse"
  /customers/me:
    get:
      tags:
        - Customer
      summary: "Get current authenticated customer's profile"
      description: "Returns the profile of the authenticated customer using their JWT token. Rate limit: 10 requests per minute per user."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer profile retrieved successfully"
          schema:
            $ref: "#/definitions/CustomerSuccessResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden - invalid role"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        429:
          description: "Too Many Requests"
          schema:
            $ref: "#/definitions/RateLimitResponse"
  /customers/me/tickets:
    get:
      tags:
        - Customer
      summary: "Get current customer's service tickets"
      description: "Returns the authenticated customer's service tickets with pagination and optional status filtering."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/StatusParam"
      responses:
        200:
          description: "Tickets retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                type: array
                items:
                  $ref: "#/definitions/CreateServiceTicketResponse"
              meta:
                $ref: "#/definitions/PaginationMeta"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
  /customers/{customer_id}:
    patch:
      tags:
        - Customer
      summary: "Update customer's basic information"
      description: "Updates name, email, or phone of the authenticated customer. Rate limited to 20 requests per hour. Only the authenticated customer can update their own profile"
      security:
        - bearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: "ID of the customer to update (must match authenticated user)"
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/CustomerSuccessResponse"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        422:
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/UnprocessableEntityResponse"
        429:
          description: "Too Many Requests"
          schema:
            $ref: "#/definitions/RateLimitResponse"
...
  /employees/login:
    post:
      tags:
        - Employee
      summary: "Employee login (generates token)"
      description: "Authenticates an employee using their employee ID and password. Returns a JWT token for subsequent requests."
      parameters:
        - in: body
          name: body
          description: "Employee Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              status: "success"
              message: "Successfully logged in"
              data:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        400:
          description: "Validation error"
          examples:
            application/json:
              errors:
                email: ["Not a valid email address.", "Invalid email format"]
              message: "Invalid input"
        401:
          description: "Invalid credentials"
          examples:
            application/json:
              message: "Invalid email or password"
              status: "error"
        403:
          description: "Account disabled"
          examples:
            application/json:
              error: "Employee account has been disabled"
        500:
          description: "Server error"
          examples:
            application/json:
              error: "Internal server error"
  /employees:
    post:
      tags:
        - Employee
      summary: "Create a new employee"
      description: |
        Creates a new employee account.
        Rate limited to 3 requests per hour.
        Password must be at least 8 characters, including letters and numbers.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: "Employee creation payload"
          required: true
          schema:
            $ref: "#/definitions/CreateEmployeePayload"
      responses:
        201:
          description: "Employee created successfully"
          schema:
            $ref: "#/definitions/CreateEmployeeResponse"
          examples:
            application/json:
              status: "success"
              message: "Employee created successfully"
              data:
                id: 1
                name: "John Smith"
                email: "john.smith@example.com"
                phone: "1234567890"
                role: "mechanic"
                salary: 60000.00
        409:
          description: "Email already exists"
          schema:
            $ref: "#/definitions/EmployeeConflictResponse"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        429:
          description: "Too Many Requests"
          produces:
            - text/html
          schema:
            type: string
            example: "<!doctype html>\n<html lang=en>\n<title>429 Too Many Requests</title>\n<h1>Too Many Requests</h1>\n<p>3 per 1 hour</p>"
          examples:
            text/html: |
              <!doctype html>
              <html lang=en>
              <title>429 Too Many Requests</title>
              <h1>Too Many Requests</h1>
              <p>3 per 1 hour</p>
        500:
          description: "Server error"
          examples:
            application/json:
              error: "Internal server error"
    
    get:
      tags:
        - Employee
      summary: "Get all employees"
      description: "Returns a paginated, filterable, and sortable list of employees. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam" 
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/NameParam"
        - $ref: "#/parameters/EmailParam"
        - $ref: "#/parameters/RoleParam"
        - $ref: "#/parameters/SearchParam"
      responses:
        200:
          description: "Employees retrieved successfully"
          schema:
            $ref: "#/definitions/GetEmployeesResponse"
        400:
          description: "Invalid query parameter"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Invalid query parameter"
          examples:
            application/json:
              status: "error"
              message: "Invalid query parameter"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
          examples:
            application/json:
              message: "Forbidden - invalid role"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /employees/customers:
    get:
      tags:
        - Employee
      summary: "Get all customers"
      description: "Admin/employee view of all customers with pagination, filtering, sorting, and search."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/NameParam"
        - $ref: "#/parameters/EmailParam"
        - $ref: "#/parameters/SearchParam"
      responses:
        200:
          description: "Customers retrieved successfully"
          schema:
            $ref: "#/definitions/GetCustomersResponse"
        400:
          description: "Invalid query parameter"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Invalid query parameter"
          examples:
            application/json:
              status: "error"
              message: "Invalid query parameter"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /employees/customers/{customer_id}:
    get:
      tags:
        - Employee
      summary: "Get customer by ID"
      description: "Returns a specific customer by ID. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: "ID of the customer to retrieve"
      responses:
        200:
          description: "Customer retrieved successfully"
          schema:
            allOf:
              - $ref: "#/definitions/BaseSuccessResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/definitions/CustomerProfileData"
          examples:
            application/json:
              status: "success"
              message: "Operation successful"
              data:
                id: 1
                name: "zambak"
                email: "zamzam@gmail.com"
                phone: "2908763453"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Customer not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Customer not found"
          examples:
            application/json:
              status: "error"
              message: "Customer not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    
    patch:
      tags:
        - Employee
      summary: "Update customer as employee"
      description: "Allows an employee to update a customer's information. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: "ID of the customer to update"
        - in: body
          name: body
          required: true
          description: "Fields to update. Leave empty to return existing profile."
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            allOf:
              - $ref: "#/definitions/BaseSuccessResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/definitions/CustomerProfileData"
          examples:
            application/json:
              status: "success"
              message: "Operation successful"
              data:
                email: "lily@gmail.com"
                id: 4
                name: "lilybun"
                phone: "5457897583"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Customer not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Customer not found"
          examples:
            application/json:
              status: "error"
              message: "Customer not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /employees/me:
    get:
      tags:
        - Employee
      summary: "Get current employee's profile"
      description: "Returns the profile of the authenticated employee. Rate limited to 10 requests per minute."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Employee profile retrieved successfully"
          schema:
            allOf:
              - $ref: "#/definitions/BaseSuccessResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/definitions/EmployeeProfileData"
          examples:
            application/json:
              status: "success"
              message: "Operation successful"
              data:
                id: 4
                name: "Jane Mechanic Queen"
                email: "janemechanic@example.com"
                phone: "1234567890"
                role: "mechanic"
                salary: "66000.00"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Employee not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Employee not found"
          examples:
            application/json:
              status: "error"
              message: "Employee not found"
        429:
          description: "Too Many Requests"
          produces:
            - text/html
          schema:
            type: string
            example: "<!doctype html>\n<html lang=en>\n<title>429 Too Many Requests</title>\n<h1>Too Many Requests</h1>\n<p>10 per 1 minute</p>"
          examples:
            text/html: |
              <!doctype html>
              <html lang=en>
              <title>429 Too Many Requests</title>
              <h1>Too Many Requests</h1>
              <p>10 per 1 minute</p>
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /employees/by-ticket-count:
    get:
      tags:
        - Employee
      summary: "Get mechanics ranked by ticket count"
      description: "Returns a list of employees ordered by the number of service tickets they're assigned to."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Employees retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    name:
                      type: string
                      example: "mike"
                    ticket_count:
                      type: integer
                      example: 3
          examples:
            application/json:
              status: "success"
              message: "Operation successful"
              data:
                - id: 1
                  name: "mike"
                  ticket_count: 3
                - id: 3
                  name: "lilo"
                  ticket_count: 2
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"

#MARK: inventory paths 
...
definitions:
...
  ValidationErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Invalid input"
      errors:
        type: object
        properties:
          field_name:
            type: array
            items:
              type: string
            example: ["Missing data for required field."]
    examples:
      application/json:
        status: "error"
        message: "Invalid input"
  UnauthorizedResponse:
    type: object
    properties:
      message:
        type: string
        example: "Authorization header is required"
    examples:
      application/json:
        message: "Authorization header is required"
...
  ServerErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Internal server error"
    examples:
      application/json:
        status: "error"
        message: "Internal server error"
...
  LoginCredentials:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string
...
  LoginSuccessResponse:
    allOf:
      - $ref: "#/definitions/BaseSuccessResponse"
      - type: object
        properties:
          data:
            type: object
            properties:
              token:
                type: string
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        examples:
          application/json:
            status: "success"
            message: "Login successful"
            data:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

#MARK: customers definitions
...